requirements.txt
Django==5.0.4
Faker==24.14.0
bcrypt==4.1.2
beautifulsoup4==4.12.3
blinker==1.7.0
django-compressor==4.4
djangorestframework-simplejwt==5.3.1
djangorestframework==3.15.1
drf-yasg==1.21.7
google-generativeai==0.6.0
googletrans==3.0.0
mongoengine==0.28.2
python-docx==1.1.2
python-dotenv==1.0.1
xhtml2pdf==0.2.15